Fix some warnings.
